pyproject.toml
[project]
name = "sysmology-cnais-mcp-server"
version = "0.1.0"
description = "A MCP server for get generative AI information about sismology from Cuba's Cenais institute"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "mcp>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "nest-asyncio>=1.6.0",
    "markdownify>=1.2.2",
]
